Move `Repeated` to separate module
